requirements.txt
aioredis==1.3.1
altgraph==0.17
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
auto-py-to-exe==2.9.0
autobahn==21.3.1
Automat==20.2.0
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
cx-Freeze==6.6
cx-Logging==3.0
cycler==0.10.0
daphne==3.0.2
Django==3.2.4
django-redis==5.0.0
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.1
importlib-metadata==4.0.1
incremental==21.3.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
msgpack==1.0.2
numpy==1.20.3
opencv-python-headless==4.5.2.54
pathlib==1.0.1
pefile==2021.5.13
Pillow==8.2.0
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
scikit-learn==0.24.2
scipy==1.6.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.1.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
whichcraft==0.6.1
whitenoise==5.2.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
